Add encoding streams suite; migrate existing TES/TDS tests
