+++ src/tests/streams/encoding/BUILD.bazel
@@ -1,0 +1,18 @@
+load("//:build/wd_test.bzl", "wd_test")
+
+# The encoding streams suite runs the same test modules under two configs:
+# encoding-cpp against the C++ implementation, encoding-ts against the
+# TypeScript implementation.
+
+encoding_suite_srcs = glob(["*.js"]) + ["encoding-modules.capnp"]
+
+wd_test(
+    src = "encoding-cpp.wd-test",
+    data = encoding_suite_srcs,
+)
+
+wd_test(
+    src = "encoding-ts.wd-test",
+    args = ["--experimental"],
+    data = encoding_suite_srcs,
+)
