--- src/workerd/api/tests/encoding-streams-test.wd-test
@@ -1,14 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  services = [
-    ( name = "encoding-streams-test",
-      worker = (
-        modules = [
-          (name = "worker", esModule = embed "encoding-streams-test.js")
-        ],
-        compatibilityFlags = ["nodejs_compat", "streams_enable_constructors", "transformstream_enable_standard_constructor"]
-      )
-    ),
-  ],
-);
